+++ ehr-web/src/app/patients/[id]/components/tabs/VitalsTab/VitalsCharts.tsx
@@ -1,0 +1,66 @@
+import React from 'react';
+import { LineChart, Line, XAxis, YAxis, CartesianGrid, Tooltip, Legend, ResponsiveContainer, Area, AreaChart } from 'recharts';
+
+interface VitalsChartsProps {
+  chartData: any[];
+}
+
+export function VitalsCharts({ chartData }: VitalsChartsProps) {
+  return (
+    <div className="grid grid-cols-2 gap-4">
+      <div className="bg-white rounded-lg border border-gray-200 p-4">
+        <h3 className="text-sm font-semibold mb-3">Blood Pressure Trend</h3>
+        <ResponsiveContainer width="100%" height={200}>
+          <LineChart data={chartData}>
+            <CartesianGrid strokeDasharray="3 3" />
+            <XAxis dataKey="date" tick={{ fontSize: 10 }} />
+            <YAxis />
+            <Tooltip />
+            <Legend />
+            <Line type="monotone" dataKey="systolic" stroke="#ef4444" strokeWidth={2} name="Systolic" />
+            <Line type="monotone" dataKey="diastolic" stroke="#3b82f6" strokeWidth={2} name="Diastolic" />
+          </LineChart>
+        </ResponsiveContainer>
+      </div>
+
+      <div className="bg-white rounded-lg border border-gray-200 p-4">
+        <h3 className="text-sm font-semibold mb-3">Heart Rate Trend</h3>
+        <ResponsiveContainer width="100%" height={200}>
+          <AreaChart data={chartData}>
+            <CartesianGrid strokeDasharray="3 3" />
+            <XAxis dataKey="date" tick={{ fontSize: 10 }} />
+            <YAxis />
+            <Tooltip />
+            <Area type="monotone" dataKey="heartRate" stroke="#10b981" fill="#10b98120" name="Heart Rate" />
+          </AreaChart>
+        </ResponsiveContainer>
+      </div>
+
+      <div className="bg-white rounded-lg border border-gray-200 p-4">
+        <h3 className="text-sm font-semibold mb-3">Temperature Trend</h3>
+        <ResponsiveContainer width="100%" height={200}>
+          <LineChart data={chartData}>
+            <CartesianGrid strokeDasharray="3 3" />
+            <XAxis dataKey="date" tick={{ fontSize: 10 }} />
+            <YAxis domain={[35, 40]} />
+            <Tooltip />
+            <Line type="monotone" dataKey="temperature" stroke="#f59e0b" strokeWidth={2} name="Temperature" />
+          </LineChart>
+        </ResponsiveContainer>
+      </div>
+
+      <div className="bg-white rounded-lg border border-gray-200 p-4">
+        <h3 className="text-sm font-semibold mb-3">O2 Saturation Trend</h3>
+        <ResponsiveContainer width="100%" height={200}>
+          <AreaChart data={chartData}>
+            <CartesianGrid strokeDasharray="3 3" />
+            <XAxis dataKey="date" tick={{ fontSize: 10 }} />
+            <YAxis domain={[90, 100]} />
+            <Tooltip />
+            <Area type="monotone" dataKey="o2sat" stroke="#6366f1" fill="#6366f120" name="O2 Saturation" />
+          </AreaChart>
+        </ResponsiveContainer>
+      </div>
+    </div>
+  );
+}
