+++ ehr-web/src/components/appointments/appointment-form-components/DrawerFooter.tsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import { Button } from '@/components/ui/button';
+
+interface DrawerFooterProps {
+  loading: boolean;
+  onCancel: () => void;
+}
+
+export function DrawerFooter({ loading, onCancel }: DrawerFooterProps) {
+  return (
+    <div className="flex items-center justify-end gap-3 border-t border-gray-200 px-6 py-4">
+      <Button
+        type="button"
+        variant="outline"
+        onClick={onCancel}
+        disabled={loading}
+      >
+        CANCEL
+      </Button>
+      <Button
+        type="submit"
+        disabled={loading}
+        className="bg-blue-600 hover:bg-blue-700"
+      >
+        {loading ? 'SAVING...' : 'SAVE'}
+      </Button>
+    </div>
+  );
+}
